+++ src/jsx-runtime.ts
@@ -1,5 +1,6 @@
-import type { Children, ComponentFn, PropBase, QNode } from "./lib/core/component.ts";
+import type { ComponentFn, PropBase, QNode } from "./lib/core/component.ts";
 import type { Tag, TagAttribute } from "./lib/core/element.ts";
+export { Fragment } from "./lib/core/component.ts"
 
 export namespace JSX {
     export interface IntrinsicElements extends IntrinsicElements_ {}
@@ -13,7 +14,7 @@ export type Component<T extends PropBase> = string | ComponentFn<T>;
 
 export type IntrinsicElements_ = { [key in keyof TagAttribute]: Partial<TagAttribute[key]> };
 
-export function jsx<T extends PropBase>(elem: Component<T>, props: T & { children?: Children }): QNode {
+export function jsx<T extends PropBase>(elem: Component<T>, props: T): QNode {
     if (typeof elem === "string") {
         return {
             tag: elem as Tag,